requirements.txt
absl-py==2.1.0
astunparse==1.6.3
black==24.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flake8==7.0.0
flatbuffers==1.12
fonttools==4.50.0
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.3.2
isort==5.13.2
Jinja2==3.1.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.1
pyflakes==3.2.0
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
typing-extensions==4.10.0
urllib3==2.2.1
werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.16.0
zipp==3.18.1
